Rewrite exercise to account for docker
Basically since we're using docker now and trying to get to a point
where psql is not installed on the host laptop at all, the exercise
needs to be re-written to account for some lost of functionality.

The major difference is that we can't use the psql meta commands like
\c and \dt. The exercise was re-written to defer that to how the GUI
app itself rather than through code. I've also split the exercise
two files, one to setup the db, then once the user switches to the
right DB, the usual todo_app.sql file that contains the rest of the
queries.